.github/workflows/docker/mqtt_endpoint/mqtt_endpoint/pyproject.toml
[project]
name = "mqtt-endpoint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=2.1.0",
    "protobuf<=3.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
endpoint = "mqtt_endpoint.main:main"
remote_control_test = "mqtt_endpoint.remote_control_test:main"
